item-1-poc-telegram-simple-chat-with-flask-v1/requirements.txt
flask=2.2.2
requests=2.31.0
python-telegram-bot=20.4
python-dotenv=1.0.0
asgiref=3.7.2